+++ src/main/kotlin/br/com/tagliaferrodev/samplerest/domain/dto/enderecos/MunicipiosPorEstadoDTO.kt
@@ -1,4 +1,4 @@
-package br.com.tagliaferrodev.samplerest.domain.dto
+package br.com.tagliaferrodev.samplerest.domain.dto.enderecos
 
 data class MunicipiosPorEstadoDTO(
         val estado: EstadoDTO,
